Refactor checkout to `useDebounce`
Move hook to global utilities.
